+++ app/src/main/java/com/example/emafoods/feature/game/domain/usecase/IncreaseXpUseCase.kt
@@ -15,7 +15,7 @@ class IncreaseXpUseCase @Inject constructor(
     private val nextLevelUseCase: GetNextLevelUseCase
 ) {
 
-    suspend fun execute(increaseXpActionType: IncreaseXpActionType, nrOfRewards: Int = 0): IncreaseXpResult<Int> {
+    suspend fun execute(increaseXpActionType: IncreaseXpActionType, nrOfRewards: Int = 0): IncreaseXpResult<Long> {
         val xpToBeAdded = if(nrOfRewards != 0) nrOfRewards * increaseXpActionType.xp else increaseXpActionType.xp
         storeUnspentUserXpUseCase.execute(xpToBeAdded)
         val unspentUserXp = getUnspentUserXpUseCase.execute()
